refactor: flatten datahub_client package structure
