+++ datahub_client/client.py
@@ -10,16 +10,6 @@ from datahub.ingestion.graph.client import DatahubClientConfig, DataHubGraph
 from datahub.metadata import schema_classes
 from datahub.metadata.schema_classes import ChangeTypeClass, DomainPropertiesClass
 
-from datahub_client.client.exceptions import ConnectivityError, EntityDoesNotExist
-from datahub_client.client.parsers import (
-    ChartParser,
-    DashboardParser,
-    DatabaseParser,
-    PublicationCollectionParser,
-    PublicationDatasetParser,
-    TableParser,
-)
-from datahub_client.client.search.search_client import SearchClient
 from datahub_client.entities import (
     Chart,
     ChartEntityMapping,
@@ -35,7 +25,17 @@ from datahub_client.entities import (
     Table,
     TableEntityMapping,
 )
-from datahub_client.search_types import (
+from datahub_client.exceptions import ConnectivityError, EntityDoesNotExist
+from datahub_client.parsers import (
+    ChartParser,
+    DashboardParser,
+    DatabaseParser,
+    PublicationCollectionParser,
+    PublicationDatasetParser,
+    TableParser,
+)
+from datahub_client.search.search_client import SearchClient
+from datahub_client.search.search_types import (
     MultiSelectFilter,
     SearchResponse,
     SortOption,
@@ -105,22 +105,22 @@ class DataHubCatalogueClient:
         self.search_client = SearchClient(self.graph)
 
         self.database_query = (
-            files("datahub_client.client.graphql")
+            files("datahub_client.graphql")
             .joinpath("getContainerDetails.graphql")
             .read_text()
         )
         self.dataset_query = (
-            files("datahub_client.client.graphql")
+            files("datahub_client.graphql")
             .joinpath("getDatasetDetails.graphql")
             .read_text()
         )
         self.chart_query = (
-            files("datahub_client.client.graphql")
+            files("datahub_client.graphql")
             .joinpath("getChartDetails.graphql")
             .read_text()
         )
         self.dashboard_query = (
-            files("datahub_client.client.graphql")
+            files("datahub_client.graphql")
             .joinpath("getDashboardDetails.graphql")
             .read_text()
         )
