+++ datahub_client/graphql/getChartDetails.graphql
