+++ datahub_client/graphql/getContainerDetails.graphql
