+++ datahub_client/graphql/getDashboardDetails.graphql
