+++ datahub_client/graphql/getDatasetDetails.graphql
