+++ datahub_client/graphql/listDomains.graphql
