+++ datahub_client/parsers.py
@@ -36,7 +36,7 @@ from datahub_client.entities import (
     TagRef,
     UsageRestrictions,
 )
-from datahub_client.search_types import SearchResult
+from datahub_client.search.search_types import SearchResult
 
 logger = logging.getLogger(__name__)
