+++ tests/home/service/test_glossary.py
@@ -1,5 +1,5 @@
 from datahub_client.entities import GlossaryTermEntityMapping
-from datahub_client.search_types import SearchResult
+from datahub_client.search.search_types import SearchResult
 from home.service.glossary import GlossaryService
